+++ src/assets/icons/gem.png
