+++ src/components/slider/SliderItem.jsx
@@ -1,6 +1,6 @@
 // SliderItem.js
 import React from "react";
-import Button from "../common/Button";
+import GradientButton from "../common/GradientButton";
 import AdInfo from "./AdInfo";
 import TributeInfo from "./TributeInfo";
 
@@ -12,7 +12,7 @@ const SliderItem = ({ idol }) => (
         src={idol.profilePicture || "defaultImage.jpg"}
         alt={idol.name}
       />
-      <Button className="tributeButton">후원하기</Button>
+      <GradientButton className="tributeButton">후원하기</GradientButton>
     </div>
     <AdInfo adLocation={idol.adLocation} name={idol.name} />
     <TributeInfo />
