+++ src/components/slider/tributeSlider.scss
@@ -68,7 +68,7 @@
   padding: 2px 16px 3px 16px;
   border-radius: 3px;
   position: absolute;
-  top: 233px; /* 이미지의 하단에 배치 */
+  bottom: 10%; /* 이미지의 하단에 배치 */
   left: 50%; /* 수평 중앙 정렬 */
   transform: translateX(-50%);
   background: linear-gradient(90deg, #f86f65 0%, $color-brand-pink 100%);
