+++ src/styles/global.scss
@@ -1,7 +1,7 @@
 @import "./font";
 @import "./color";
 @import "../components/slider/tributeSlider.scss";
-
+@import "../components/common/GradientButton.scss";
 
 *,
 *::before,
